refactor courses components
